Fix Android build: Kotlin + CMake + install on device
Kotlin fixes:
- Add AlertDialog import (MainActivity.kt)
- Add kotlin.math.min import (MicStreamingService.kt)
- Add android.content.Context import (MicStreamingService.kt)
- Move opusEncoder to outer scope for finally-block visibility
- Fix Byte.or() → Byte.toInt().or().toByte() (PacketCrypto.kt)
- Add @JvmStatic to JNI declarations (OpusEncoder.kt)
- Match library name: pocketmic_opus → pocketmic_jni

CMake fixes:
- opus_jni.c: #include <opus.h> not <opus/opus.h>
- opus_jni.c: OPUS_SET_INBAND_FEC not OPUS_SET_FEC

Build result:
- assembleDebug: SUCCESS
- APK: 35MB, all 3 ABIs (arm64-v8a, armeabi-v7a, x86_64)
- libopus.so + libpocketmic_jni.so in all ABIs
- Install on OnePlus 9 Pro (Android 14): SUCCESS

+++ android/app/src/main/java/com/ryanspice/pocketmic/PacketCrypto.kt
@@ -69,7 +69,7 @@ object PacketCrypto {
             .order(ByteOrder.BIG_ENDIAN)
             .put(MAGIC)
             .put(VERSION_V2)
-            .put((FLAG_ENCRYPTED or FLAG_OPUS).toByte())
+            .put((FLAG_ENCRYPTED.toInt() or FLAG_OPUS.toInt()).toByte())
             .putShort(HEADER_SIZE_V2.toShort())
             .putLong(sessionId)
             .putInt(sequence)
@@ -199,7 +199,7 @@ object PacketCrypto {
                     // Write v2 header
                     MAGIC.copyInto(packet, destinationOffset = 0)
                     packet[4] = VERSION_V2
-                    packet[5] = (FLAG_ENCRYPTED or FLAG_OPUS).toByte()
+                    packet[5] = (FLAG_ENCRYPTED.toInt() or FLAG_OPUS.toInt()).toByte()
                     writeShortBigEndian(packet, 6, HEADER_SIZE_V2)
                     writeLongBigEndian(packet, 8, sessionId)
                     writeIntBigEndian(packet, 16, sequence)
